:art: Rename: Authentication package structure
