+++ spring-boot-api/src/main/java/com/ssafy/springbootapi/global/auth/jsonAuthentication/JsonUserAuthenticationSuccessHandler.java
@@ -1,19 +1,14 @@
-package com.ssafy.springbootapi.global.auth;
+package com.ssafy.springbootapi.global.auth.jsonAuthentication;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.ssafy.springbootapi.domain.user.application.UserService;
-import com.ssafy.springbootapi.domain.user.dao.UserRepository;
-import com.ssafy.springbootapi.domain.user.domain.User;
 import com.ssafy.springbootapi.domain.user.dto.UserLoginResponse;
 import com.ssafy.springbootapi.global.auth.jwt.TokenProvider;
-import com.ssafy.springbootapi.global.auth.jwt.refreshToken.RefreshToeknRepository;
 import com.ssafy.springbootapi.global.auth.jwt.refreshToken.RefreshToken;
+import com.ssafy.springbootapi.global.auth.jwt.refreshToken.RefreshTokenRepository;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
 
@@ -21,16 +16,14 @@ import java.io.IOException;
 import java.time.Duration;
 
 @Component
-public class MyAuthenticationSuccessHandler implements AuthenticationSuccessHandler {
+public class JsonUserAuthenticationSuccessHandler implements AuthenticationSuccessHandler {
 
     private final TokenProvider tokenProvider;
-    private final RefreshToeknRepository refreshToeknRepository;
-    private final UserRepository userRepository;
+    private final RefreshTokenRepository refreshTokenRepository;
 
-    public MyAuthenticationSuccessHandler(TokenProvider tokenProvider, RefreshToeknRepository refreshToeknRepository, UserRepository userRepository) {
+    public JsonUserAuthenticationSuccessHandler(TokenProvider tokenProvider, RefreshTokenRepository refreshTokenRepository) {
         this.tokenProvider = tokenProvider;
-        this.refreshToeknRepository = refreshToeknRepository;
-        this.userRepository = userRepository;
+        this.refreshTokenRepository = refreshTokenRepository;
     }
 
     @Override
@@ -41,12 +34,14 @@ public class MyAuthenticationSuccessHandler implements AuthenticationSuccessHand
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
         // 여기에서 UserService의 login을 실행
-        User user = userRepository.findByEmail(authentication.getName()).orElseThrow();
-        String accessToken = tokenProvider.generateToken(user, Duration.ofMinutes(1L));
-        String refreshToken = tokenProvider.generateToken(user, Duration.ofMinutes(30L));
+//        User user = userRepository.findByEmail(authentication.getName()).orElseThrow();
+        String email = authentication.getName();
+
+        String accessToken = tokenProvider.generateToken(email, Duration.ofMinutes(1L));
+        String refreshToken = tokenProvider.generateToken(email, Duration.ofMinutes(30L));
         System.out.println("accessToken "+accessToken);
-        refreshToeknRepository.save(RefreshToken.builder()
-                .userId(user.getId())
+        refreshTokenRepository.save(RefreshToken.builder()
+                .email(email)
                 .refreshToken(refreshToken)
                 .build());
         response.setContentType("application/json");
