+++ spring-boot-api/src/main/java/com/ssafy/springbootapi/global/auth/jsonAuthentication/JsonUsernamePasswordAuthenticationFilter.java
@@ -1,11 +1,9 @@
-package com.ssafy.springbootapi.global.auth;
+package com.ssafy.springbootapi.global.auth.jsonAuthentication;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ssafy.springbootapi.domain.user.dto.UserLoginRequest;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
